[Refactor] Moved some io methods to public API
