+++ src/occa/internal/io/utils.cpp
@@ -338,16 +338,6 @@ namespace occa {
       return filesInDir(dir, DT_REG);
     }
 
-    bool exists(const std::string &filename) {
-      std::string expFilename = io::filename(filename);
-      FILE *fp = fopen(expFilename.c_str(), "rb");
-      if (fp == NULL) {
-        return false;
-      }
-      fclose(fp);
-      return true;
-    }
-
     char* c_read(const std::string &filename,
                  size_t *chars,
                  enums::FileType fileType) {
